refactor folders for discovery integration tests
