+++ modules/outbound-command-event-handler/test/integration/application/discovery/process_bulk_accept_party_info.test.ts
